migrated app to use browserify
