+++ .gitignore
@@ -1,3 +1,5 @@
 bower_components
+node_modules
 .DS_Store
 .idea
+static/js/bundle-debug.js
