+++ app/js/app.js
@@ -1,13 +1,17 @@
 'use strict';
 
-/* global UI, API, Map, FbUser */
-
-// Markers are currently stored in an array in Map module
+// Markers are currently stored in an array in GMap module
 // and the data retrieved from API calls is stored in customInfo
 // property of each marker. Once a marker is clicked, customInfo property is
 // passed into the UI and stored. UI actions handlers are implemented below,
 // they revceive the customInfo object as parameter and can modify it
 // here if needed and notify any changes to the API or MAP objects.
+
+require('babelify/polyfill');
+var API = require('./apiclient.js').API;
+var UI = require('./ui').UI;
+var GMap = require('./map').Map;
+var FbUser = require('./fblogin').FbUser;
 
 window.addEventListener('DOMContentLoaded', function() {
   if(localStorage.getItem('debug') === 'enabled') {
@@ -15,7 +19,7 @@ window.addEventListener('DOMContentLoaded', function() {
   }
 
   var addMarker = function(markerOptions) {
-    var marker = Map.createMarker(markerOptions);
+    var marker = GMap.createMarker(markerOptions);
     API.addMarker(marker.customInfo, function(savedMarker) {
       marker.customInfo.id = savedMarker.id;
     });
@@ -23,7 +27,7 @@ window.addEventListener('DOMContentLoaded', function() {
 
   var deleteMarker = function(marker) {
     API.deleteMarker(marker, function() {
-      Map.removeMarker(marker);
+      GMap.removeMarker(marker);
     });
   };
 
@@ -37,7 +41,7 @@ window.addEventListener('DOMContentLoaded', function() {
     marker.state = 'failure';
     API.updateMarker(marker, function() {
       UI.showDetails(marker);
-      Map.redrawMarker(marker);
+      GMap.redrawMarker(marker);
     });
   };
 
@@ -75,12 +79,12 @@ window.addEventListener('DOMContentLoaded', function() {
     UI.showDetails(marker);
   };
 
-  Map.setMarkerMovedHandler(markerMoved);
-  Map.setMarkerSelectedHandler(markerSelected);
+  GMap.setMarkerMovedHandler(markerMoved);
+  GMap.setMarkerSelectedHandler(markerSelected);
 
-  Map.init();
+  GMap.init();
   API.getMarkers(null, function(markers) {
-    Map.drawMarkers(markers);
+    GMap.drawMarkers(markers);
   });
 
 });
