+++ index.html
@@ -6,11 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
     <title>EasyCity</title>
-
-    <!-- Bootstrap -->
-    <link href="static/css/bootstrap.css" rel="stylesheet">
-    <link href="static/css/lato-font.css" rel="stylesheet">
-    <link href="static/css/app.css" rel="stylesheet">
+    <link href="static/css/style.css" rel="stylesheet">
 
     <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
     <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
@@ -20,8 +16,6 @@
     <![endif]-->
   </head>
   <body>
-    <section id="welcome"></section>
-    <section id="categories"></section>
     <section id="main">
       <nav class="navbar navbar-default navbar-fixed-top">
         <div class="container">
@@ -142,25 +136,7 @@
         </div>
       </section>
     </section>
-    <script src="static/js/config.js"></script>
-    <script src="http://maps.google.com/maps/api/js?sensor=false"></script>
-    <script src="static/js/fblogin.js"></script>
-    <script src="static/js/libs/jquery.js"></script>
-    <script src="static/js/libs/bootstrap.js"></script>
-    <script src="static/js/libs/geolocationmarker-compiled.js"></script>
-    <script src="static/js/polyfills.js"></script>
-    <script src="static/js/apiclient.js"></script>
-    <script src="static/js/ui.js"></script>
-    <script src="static/js/map.js"></script>
-    <script src="static/js/app.js"></script>
-    <!-- Google Tag Manager -->
-    <noscript><iframe src="//www.googletagmanager.com/ns.html?id=GTM-NG48BL"
-    height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
-    <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-    new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-    j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-    '//www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-    })(window,document,'script','dataLayer','GTM-NG48BL');</script>
-    <!-- End Google Tag Manager -->
+    <script type="text/javascript" src="http://maps.google.com/maps/api/js?sensor=false"></script>
+    <script type="text/javascript" src="static/js/bundle-debug.js"></script>
   </body>
 </html>
