--- static/css/app.css
@@ -1,228 +1,0 @@
-html {
-  font-size: 20px;
-}
-
-body {
-  font-family: "Lato";
-  font-color: white;
-  overflow: hidden;
-}
-
-.navbar {
-  background-color: #58bac3;
-  color: white;
-  border: none;
-  height: 2rem;
-  padding-top: 0.2rem;
-}
-
-.navbar .navbar-text {
-  color: white;
-  font-weight: bold;
-}
-
-.main-title {
-  background-image: url(../img/logo.png);
-  width: 115px;
-  height: 35px;
-  background-size: contain;
-  background-repeat: no-repeat;
-  margin-top: 0.2rem;
-  margin-left: 1rem;
-}
-
-button, .icon {
-  border: none;
-  width: 2rem;
-  height: 2rem;
-  outline: none;
-  padding: 0.2rem;
-  background-position: center;
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-color: inherit;
-}
-
-.btn-big, .icon-big {
-  width: 4rem;
-  height: 4rem;
-  background-color: white;
-}
-
-#main-button-back {
-  background-image: url(../img/icon/back.png);
-  background-color: #58bac3;
-}
-
-#main-button-add {
-  background-image: url(../img/icon/add.png);
-  background-color: #58bac3;
-}
-
-#btn-type-select-facility {
-  background-image: url(../img/icon/udogodnienia.png);
-}
-
-#btn-type-select-obstacle {
-  background-image: url(../img/icon/przeszkody.png);
-}
-
-#btn-type-select-failure {
-  background-image: url(../img/icon/awaria.png);
-}
-
-.icon-obs-nierownosci {
-  background-image: url(../img/icon/nierownosci.png);
-}
-
-.icon-obs-pochylosc {
-  background-image: url(../img/icon/pochylosc.png);
-}
-
-.icon-obs-prznad {
-  background-image: url(../img/icon/przejscienadziemne.png);
-}
-
-.icon-obs-schody {
-  background-image: url(../img/icon/schody.png);
-}
-
-.icon-obs-kraweznik {
-  background-image: url(../img/icon/wysokikraweznik.png);
-}
-
-.icon-fac-poch {
-  background-image: url(../img/icon/pochylnia.png);
-}
-
-.icon-fac-pod {
-  background-image: url(../img/icon/podnosnik.png);
-}
-
-.icon-fac-kraw {
-  background-image: url(../img/icon/przyjaznykraweznik.png);
-}
-
-.icon-fac-winda {
-  background-image: url(../img/icon/winda.png);
-}
-
-.icon-action-przeszkadzami {
-  background-image: url(../img/icon/przeszkadzmi.png)
-}
-
-.icon-action-awaria {
-  background-image: url(../img/icon/awaria.png);
-}
-
-section#main-map {
-  height: 100vh;
-  margin-top: 50px;
-}
-
-.action > section#main-map {
-  height: 40vh;
-}
-
-section.action-section {
-  height: 60vh;
-  background-color: white;
-  position: relative;
-}
-
-.action-add > section#main-add {
-  z-index: 10;
-}
-
-.action-details > section#main-details {
-  top: -60vh;
-  z-index: 20;
-}
-
-.action-add > section#main-details {
-  display: none;
-}
-
-.action-section header {
-  text-align: center;
-  height: 2rem;
-  line-height: 2rem;
-  background-color: #58bac3;
-  color: white;
-  font-size: 1.1rem;
-}
-
-.type-select {
-  text-align: center;
-  margin-top: 1rem;
-}
-
-.select-facility #main-add-type-select,
-.select-obstacle #main-add-type-select,
-.select-failure #main-add-type-select {
-  display: none;
-}
-
-.select-facility #main-add-obstacle,
-.select-failure #main-add-obstacle,
-.select-type #main-add-obstacle {
-  display: none;
-}
-
-.select-facility #main-add-failure,
-.select-obstacle #main-add-failure,
-.select-type #main-add-failure {
-  display: none;
-}
-
-.select-obstacle #main-add-facility,
-.select-failure #main-add-facility,
-.select-type #main-add-facility {
-  display: none;
-}
-
-div.col-xs-4, div.col-xs-12 {
-  text-align: center;
-}
-
-div.col-xs-4 > div{
-  display: inline-block;
-}
-
-#main-details .container {
-  margin-top: 1rem;
-}
-
-#main-details-category {
-  font-size: 1.2rem;
-}
-
-.error {
-  color: red;
-}
-
-#main-details-upvotes {
-  font-style: italic;
-}
-
-.main-details-actions {
-  margin-top: 1rem;
-  border-top: 0.2rem dotted #58bac3;
-  padding: 0.5rem;
-}
-
-.type-obstacle .awaria {
-  display: none;
-}
-
-.type-facility #main-details-upvotes {
-  display: none;
-}
-
-.type-facility .przeszkadzami {
-  display: none;
-}
-
-.hide {
-  display: none;
-}
