--- static/css/map.css
@@ -1,4 +1,0 @@
-#mapCanvas {
-    width: 100%;
-    height: 500px;
-}
